requirements.txt
PyYAML==6.0.2
transformers==4.44.1
